+++ rules/imports.js
@@ -5,5 +5,6 @@ module.exports = {
         'import/extensions': 'off',
         'import/no-relative-parent-imports': 'off',
         'import/no-extraneous-dependencies': 'off',
+        'import/prefer-default-export': 'off',
     },
 };