+++ rules/typescript.js
@@ -1,0 +1,8 @@
+module.exports = {
+    rules: {
+        '@typescript-eslint/ban-ts-comment': 'off',
+        '@typescript-eslint/no-explicit-any': 'off',
+        '@typescript-eslint/no-var-requires': 'off',
+        '@typescript-eslint/interface-name-prefix': 'off',
+    },
+};