+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.0] - 2021-11-19
+### Added
+- TypeScript support via extending `@kellerkinder/eslint-config/typescript`
+- TypeScript specific rules based on `@typescript-eslint/recommended`
+
+### Changed
+- `import/prefer-default-export` rule disabled
+
 ## [2.0.0] - 2021-08-11
 ### Added
 - Default babel configuration
@@ -63,7 +71,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Basic implementation based on AirBnB ESLint configuration
 
-[Unreleased]: https://github.com/kellerkinderDE/eslint-config/compare/2.0.0...HEAD
+[Unreleased]: https://github.com/kellerkinderDE/eslint-config/compare/2.1.0...HEAD
+[2.1.0]: https://github.com/kellerkinderDE/eslint-config/compare/2.0.0...2.1.0
 [2.0.0]: https://github.com/kellerkinderDE/eslint-config/compare/1.1.0...2.0.0
 [1.1.0]: https://github.com/kellerkinderDE/eslint-config/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/kellerkinderDE/eslint-config/compare/1.0.0-rc.4...1.0.0